+++ apps/common/ospapp/OSPApp.h
@@ -66,6 +66,7 @@ namespace ospray {
       void addGeneratorNodesToWorld(sg::Node &renderer);
       void addAnimatedImporterNodesToWorld(sg::Node &renderer);
       void setupCamera(sg::Node &renderer);
+      void setupToneMapping(sg::Node &renderer);
       void addPlaneToScene(sg::Node &renderer);
       void printHelp();
 
@@ -106,6 +107,8 @@ namespace ospray {
       std::string hdriLightFile;
       bool addDefaultLights = false;
       bool noDefaultLights = false;
+      bool aces = false;
+      bool filmic = false;
       bool debug = false;
       std::string initialRendererType;
       box3f bboxWithoutPlane;