+++ components/ospcommon/vec.h
@@ -33,6 +33,9 @@ namespace ospcommon {
   namespace traits {
 
     template <typename T>
+    using is_arithmetic_t = enable_if_t<std::is_arithmetic<T>::value>;
+
+    template <typename T>
     struct is_vec
     {
       const static bool value = std::is_base_of<vec_base, T>::value;
@@ -385,7 +388,7 @@ namespace ospcommon {
 #undef unary_functor
 
 // -------------------------------------------------------
-// binary operators, same type
+// binary arithmetic operators
 // -------------------------------------------------------
 
 # define binary_operator(name, op)                                             \
@@ -425,52 +428,64 @@ namespace ospcommon {
   template <typename T,                                                        \
             typename U = T,                                                    \
             typename = traits::is_valid_scalar_for_binary_op_t<vec_t<T, 2>, U>>\
-  inline vec_t<T, 2> name(const vec_t<T, 2> &a, const U &b)                    \
+  inline auto name(const vec_t<T, 2> &a, const U &b)                           \
+      -> vec_t<decltype(T() op U()), 2>                                        \
   {                                                                            \
-    return vec_t<T, 2>(a.x op T(b), a.y op T(b));                              \
+    using result_t = vec_t<decltype(T() op U()), 2>;                           \
+    return result_t(a.x op b, a.y op b);                                       \
   }                                                                            \
                                                                                \
   template <typename T,                                                        \
             bool A,                                                            \
             typename U = T,                                                    \
             typename = traits::is_valid_scalar_for_binary_op_t<vec_t<T, 3, A>, U>> \
-  inline vec_t<T, 3, A> name(const vec_t<T, 3, A> &a, const U &b)              \
+  inline auto name(const vec_t<T, 3, A> &a, const U &b)                        \
+      -> vec_t<decltype(T() op U()), 3, A>                                     \
   {                                                                            \
-    return vec_t<T, 3, A>(a.x op T(b), a.y op T(b), a.z op T(b));              \
+    using result_t = vec_t<decltype(T() op U()), 3, A>;                        \
+    return result_t(a.x op b, a.y op b, a.z op b);                             \
   }                                                                            \
                                                                                \
   template <typename T,                                                        \
             typename U = T,                                                    \
             typename = traits::is_valid_scalar_for_binary_op_t<vec_t<T, 4>, U>>\
-  inline vec_t<T, 4> name(const vec_t<T, 4> &a, const U &b)                    \
+  inline auto name(const vec_t<T, 4> &a, const U &b)                           \
+      -> vec_t<decltype(T() op U()), 4>                                        \
   {                                                                            \
-    return vec_t<T, 4>(a.x op T(b), a.y op T(b), a.z op T(b), a.w op T(b));    \
+    using result_t = vec_t<decltype(T() op U()), 4>;                           \
+    return result_t(a.x op b, a.y op b, a.z op b, a.w op b);                   \
   }                                                                            \
                                                                                \
   /* "scalar op vec" (SFINAE out scalar types which are ill-formed) */         \
   template <typename T,                                                        \
             typename U = T,                                                     \
             typename = traits::is_valid_scalar_for_binary_op_t<vec_t<T, 2>, U>> \
-  inline vec_t<T, 2> name(const U &a, const vec_t<T, 2> &b)                    \
+  inline auto name(const U &a, const vec_t<T, 2> &b)                           \
+      -> vec_t<decltype(U() op T()), 2>                                        \
   {                                                                            \
-    return vec_t<T, 2>(T(a) op b.x, T(a) op b.y);                              \
+    using result_t = vec_t<decltype(U() op T()), 2>;                           \
+    return result_t(a op b.x, a op b.y);                                       \
   }                                                                            \
                                                                                \
   template <typename T,                                                        \
             bool A,                                                            \
             typename U = T,                                                    \
             typename = traits::is_valid_scalar_for_binary_op_t<vec_t<T, 3, A>, U>> \
-  inline vec_t<T, 3, A> name(const U &a, const vec_t<T, 3, A> &b)              \
+  inline auto name(const U &a, const vec_t<T, 3, A> &b)                        \
+      -> vec_t<decltype(U() op T()), 3, A>                                     \
   {                                                                            \
-    return vec_t<T, 3, A>(T(a) op b.x, a op b.y, T(a) op b.z);                 \
+    using result_t = vec_t<decltype(U() op T()), 3, A>;                        \
+    return result_t(a op b.x, a op b.y, a op b.z);                             \
   }                                                                            \
                                                                                \
   template <typename T,                                                        \
             typename U = T,                                                    \
             typename = traits::is_valid_scalar_for_binary_op_t<vec_t<T, 4>, U>>\
-  inline vec_t<T, 4> name(const U &a, const vec_t<T, 4> &b)                    \
+  inline auto name(const U &a, const vec_t<T, 4> &b)                           \
+      -> vec_t<decltype(U() op T()), 4>                                        \
   {                                                                            \
-    return vec_t<T, 4>(T(a) op b.x, T(a) op b.y, T(a) op b.z, T(a) op b.w);    \
+    using result_t = vec_t<decltype(U() op T()), 4>;                           \
+    return result_t(a op b.x, a op b.y, a op b.z, a op b.w);                   \
   }
 
   binary_operator(operator+, +)
@@ -481,20 +496,20 @@ namespace ospcommon {
 #undef binary_operator
 
 // -------------------------------------------------------
-// binary operators, same type
+// binary arithmetic assignment operators
 // -------------------------------------------------------
 #define binary_operator(name, op)                                         \
   /* "vec op vec" */                                                      \
-  template <typename T>                                                   \
-  inline vec_t<T, 2> &name(vec_t<T, 2> &a, const vec_t<T, 2> &b)          \
+  template <typename T, typename U>                                       \
+  inline vec_t<T, 2> &name(vec_t<T, 2> &a, const vec_t<U, 2> &b)          \
   {                                                                       \
     a.x op b.x;                                                           \
     a.y op b.y;                                                           \
     return a;                                                             \
   }                                                                       \
                                                                           \
-  template <typename T, bool A, bool B>                                   \
-  inline vec_t<T, 3, A> &name(vec_t<T, 3, A> &a, const vec_t<T, 3, B> &b) \
+  template <typename T, typename U, bool A, bool B>                       \
+  inline vec_t<T, 3, A> &name(vec_t<T, 3, A> &a, const vec_t<U, 3, B> &b) \
   {                                                                       \
     a.x op b.x;                                                           \
     a.y op b.y;                                                           \
@@ -502,8 +517,8 @@ namespace ospcommon {
     return a;                                                             \
   }                                                                       \
                                                                           \
-  template <typename T>                                                   \
-  inline vec_t<T, 4> &name(vec_t<T, 4> &a, const vec_t<T, 4> &b)          \
+  template <typename T, typename U>                                       \
+  inline vec_t<T, 4> &name(vec_t<T, 4> &a, const vec_t<U, 4> &b)          \
   {                                                                       \
     a.x op b.x;                                                           \
     a.y op b.y;                                                           \
@@ -513,16 +528,21 @@ namespace ospcommon {
   }                                                                       \
                                                                           \
   /* "vec op scalar" */                                                   \
-  template <typename T>                                                   \
-  inline vec_t<T, 2> &name(vec_t<T, 2> &a, const T &b)                    \
+  template <typename T,                                                   \
+            typename U,                                                   \
+            typename = traits::is_arithmetic_t<U>>                        \
+  inline vec_t<T, 2> &name(vec_t<T, 2> &a, const U &b)                    \
   {                                                                       \
     a.x op b;                                                             \
     a.y op b;                                                             \
     return a;                                                             \
   }                                                                       \
                                                                           \
-  template <typename T, bool A>                                           \
-  inline vec_t<T, 3, A> &name(vec_t<T, 3, A> &a, const T &b)              \
+  template <typename T,                                                   \
+            typename U,                                                   \
+            bool A,                                                       \
+            typename = traits::is_arithmetic_t<U>>                        \
+  inline vec_t<T, 3, A> &name(vec_t<T, 3, A> &a, const U &b)              \
   {                                                                       \
     a.x op b;                                                             \
     a.y op b;                                                             \
@@ -530,8 +550,10 @@ namespace ospcommon {
     return a;                                                             \
   }                                                                       \
                                                                           \
-  template <typename T>                                                   \
-  inline vec_t<T, 4> &name(vec_t<T, 4> &a, const T &b)                    \
+  template <typename T,                                                   \
+            typename U,                                                   \
+            typename = traits::is_arithmetic_t<U>>                        \
+  inline vec_t<T, 4> &name(vec_t<T, 4> &a, const U &b)                    \
   {                                                                       \
     a.x op b;                                                             \
     a.y op b;                                                             \