+++ ospray/render/pathtracer/materials/Principled.cpp
@@ -38,83 +38,76 @@ namespace ospray {
       //! \brief commit the material's parameters
       virtual void commit() override
       {
-        Texture2D* baseColorMap = (Texture2D*)getParamObject("baseColorMap");
-        affine2f baseColorXform = getTextureTransform("baseColorMap");
-        vec3f baseColor = getParam3f("baseColor", baseColorMap ? vec3f(1.f) : vec3f(0.8f));
+        MaterialParam3f baseColor = getMaterialParam3f("baseColor", vec3f(0.8f));
+        MaterialParam3f edgeColor = getMaterialParam3f("edgeColor", vec3f(1.f));
+        MaterialParam1f metallic = getMaterialParam1f("metallic", 0.f);
+        MaterialParam1f diffuse = getMaterialParam1f("diffuse", 1.f);
+        MaterialParam1f specular = getMaterialParam1f("specular", 1.f);
+        MaterialParam1f ior = getMaterialParam1f("ior", 1.f);
+        MaterialParam1f transmission = getMaterialParam1f("transmission", 0.f);
+        MaterialParam3f transmissionColor = getMaterialParam3f("transmissionColor", vec3f(1.f));
+        MaterialParam1f transmissionDepth = getMaterialParam1f("transmissionDepth", 1.f);
+        MaterialParam1f roughness = getMaterialParam1f("roughness", 0.f);
+        MaterialParam1f anisotropy = getMaterialParam1f("anisotropy", 0.f);
+        MaterialParam1f rotation = getMaterialParam1f("rotation", 0.f);
+        MaterialParam1f normal = getMaterialParam1f("normal", 1.f);
 
-        Texture2D* metallicMap = (Texture2D*)getParamObject("metallicMap");
-        affine2f metallicXform = getTextureTransform("metallicMap");
-        float metallic = getParamf("metallic", metallicMap ? 1.f : 0.f);
+        MaterialParam1f coat = getMaterialParam1f("coat", 0.f);
+        MaterialParam1f coatIor = getMaterialParam1f("coatIor", 1.5f);
+        MaterialParam3f coatColor = getMaterialParam3f("coatColor", vec3f(1.f));
+        MaterialParam1f coatThickness = getMaterialParam1f("coatThickness", 1.f);
+        MaterialParam1f coatRoughness = getMaterialParam1f("coatRoughness", 0.f);
+        MaterialParam1f coatNormal = getMaterialParam1f("coatNormal", 1.f);
 
-        Texture2D* specularMap = (Texture2D*)getParamObject("specularMap");
-        affine2f specularXform = getTextureTransform("specularMap");
-        float specular = getParamf("specular", specularMap ? 1.f : 0.f);
+        MaterialParam1f sheen = getMaterialParam1f("sheen", 0.f);
+        MaterialParam3f sheenColor = getMaterialParam3f("sheenColor", vec3f(1.f));
+        MaterialParam1f sheenRoughness = getMaterialParam1f("sheenRoughness", 0.2f);
 
-        Texture2D* edgeColorMap = (Texture2D*)getParamObject("edgeColorMap");
-        affine2f edgeColorXform = getTextureTransform("edgeColorMap");
-        vec3f edgeColor = getParam3f("edgeColor", vec3f(1.f));
+        MaterialParam1f opacity = getMaterialParam1f("opacity", 1.f);
 
-        Texture2D* transmissionMap = (Texture2D*)getParamObject("transmissionMap");
-        affine2f transmissionXform = getTextureTransform("transmissionMap");
-        float transmission = getParamf("transmission", transmissionMap ? 1.f : 0.f);
+        bool thin = getParam1i("thin", 0) || (getParam1f("thin", 0.f) > 0.f);
+        MaterialParam1f backlight = getMaterialParam1f("backlight", 0.f);
+        MaterialParam1f thickness = getMaterialParam1f("thickness", 1.f);
 
-        Texture2D* roughnessMap = (Texture2D*)getParamObject("roughnessMap");
-        affine2f roughnessXform = getTextureTransform("roughnessMap");
-        float roughness = getParamf("roughness", roughnessMap ? 1.f : 0.f);
-
-        Texture2D* normalMap = (Texture2D*)getParamObject("normalMap");
-        affine2f normalXform = getTextureTransform("normalMap");
-        linear2f normalRot   = normalXform.l.orthogonal().transposed();
-        float normalScale = getParamf("normalScale", 1.f);
-
-        Texture2D* coatMap = (Texture2D*)getParamObject("coatMap");
-        affine2f coatXform = getTextureTransform("coatMap");
-        float coat = getParamf("coat", coatMap ? 1.f : 0.f);
-
-        Texture2D* coatColorMap = (Texture2D*)getParamObject("coatColorMap");
-        affine2f coatColorXform = getTextureTransform("coatColorMap");
-        vec3f coatColor = getParam3f("coatColor", vec3f(1.f));
-
-        Texture2D* coatThicknessMap = (Texture2D*)getParamObject("coatThicknessMap");
-        affine2f coatThicknessXform = getTextureTransform("coatThicknessMap");
-        float coatThickness = getParamf("coatThickness", 1.f);
-
-        Texture2D* coatRoughnessMap = (Texture2D*)getParamObject("coatRoughnessMap");
-        affine2f coatRoughnessXform = getTextureTransform("coatRoughnessMap");
-        float coatRoughness = getParamf("coatRoughness", coatRoughnessMap ? 1.f : 0.f);
-
-        Texture2D* coatNormalMap = (Texture2D*)getParamObject("coatNormalMap");
-        affine2f coatNormalXform = getTextureTransform("coatNormalMap");
-        linear2f coatNormalRot   = coatNormalXform.l.orthogonal().transposed();
-        float coatNormalScale = getParamf("coatNormalScale", 1.f);
-
-        float ior = getParamf("ior", 1.5f);
-        vec3f transmissionColor = getParam3f("transmissionColor", vec3f(1.f));
-        float transmissionDepth = getParamf("transmissionDepth", 1.f);
-
-        float iorOutside = getParamf("iorOutside", 1.f);
-        vec3f transmissionColorOutside = getParam3f("transmissionColorOutside", vec3f(1.f));
-        float transmissionDepthOutside = getParamf("transmissionDepthOutside", 1.f);
+        float outsideIor = getParamf("outsideIor", 1.f);
+        vec3f outsideTransmissionColor = getParam3f("outsideTransmissionColor", vec3f(1.f));
+        float outsideTransmissionDepth = getParamf("outsideTransmissionDepth", 1.f);
 
         ispc::PathTracer_Principled_set(getIE(),
-          (const ispc::vec3f&)baseColor, baseColorMap ? baseColorMap->getIE() : nullptr, (const ispc::AffineSpace2f&)baseColorXform,
-          metallic, metallicMap ? metallicMap->getIE() : nullptr, (const ispc::AffineSpace2f&)metallicXform,
-          specular, specularMap ? specularMap->getIE() : nullptr, (const ispc::AffineSpace2f&)specularXform,
-          (const ispc::vec3f&)edgeColor, edgeColorMap ? edgeColorMap->getIE() : nullptr, (const ispc::AffineSpace2f&)edgeColorXform,
-          transmission, transmissionMap ? transmissionMap->getIE() : nullptr, (const ispc::AffineSpace2f&)transmissionXform,
-          roughness, roughnessMap ? roughnessMap->getIE() : nullptr, (const ispc::AffineSpace2f&)roughnessXform,
-          normalMap ? normalMap->getIE() : nullptr, (const ispc::AffineSpace2f&)normalXform, (const ispc::LinearSpace2f&)normalRot, normalScale,
-          coat, coatMap ? coatMap->getIE() : nullptr, (const ispc::AffineSpace2f&)coatXform,
-          (const ispc::vec3f&)coatColor, coatColorMap ? coatColorMap->getIE() : nullptr, (const ispc::AffineSpace2f&)coatColorXform,
-          coatThickness, coatThicknessMap ? coatThicknessMap->getIE() : nullptr, (const ispc::AffineSpace2f&)coatThicknessXform,
-          coatRoughness, coatRoughnessMap ? coatRoughnessMap->getIE() : nullptr, (const ispc::AffineSpace2f&)coatRoughnessXform,
-          coatNormalMap ? coatNormalMap->getIE() : nullptr, (const ispc::AffineSpace2f&)coatNormalXform, (const ispc::LinearSpace2f&)coatNormalRot, coatNormalScale,
-          ior,
-          (const ispc::vec3f&)transmissionColor,
-          transmissionDepth,
-          iorOutside,
-          (const ispc::vec3f&)transmissionColorOutside,
-          transmissionDepthOutside);
+          (const ispc::vec3f&)baseColor.factor, baseColor.map ? baseColor.map->getIE() : nullptr, (const ispc::AffineSpace2f&)baseColor.xform,
+          (const ispc::vec3f&)edgeColor.factor, edgeColor.map ? edgeColor.map->getIE() : nullptr, (const ispc::AffineSpace2f&)edgeColor.xform,
+          metallic.factor, metallic.map ? metallic.map->getIE() : nullptr, (const ispc::AffineSpace2f&)metallic.xform,
+          diffuse.factor, diffuse.map ? diffuse.map->getIE() : nullptr, (const ispc::AffineSpace2f&)diffuse.xform,
+          specular.factor, specular.map ? specular.map->getIE() : nullptr, (const ispc::AffineSpace2f&)specular.xform,
+          ior.factor, ior.map ? ior.map->getIE() : nullptr, (const ispc::AffineSpace2f&)ior.xform,
+          transmission.factor, transmission.map ? transmission.map->getIE() : nullptr, (const ispc::AffineSpace2f&)transmission.xform,
+          (const ispc::vec3f&)transmissionColor.factor, transmissionColor.map ? transmissionColor.map->getIE() : nullptr, (const ispc::AffineSpace2f&)transmissionColor.xform,
+          transmissionDepth.factor, transmissionDepth.map ? transmissionDepth.map->getIE() : nullptr, (const ispc::AffineSpace2f&)transmissionDepth.xform,
+          roughness.factor, roughness.map ? roughness.map->getIE() : nullptr, (const ispc::AffineSpace2f&)roughness.xform,
+          anisotropy.factor, anisotropy.map ? anisotropy.map->getIE() : nullptr, (const ispc::AffineSpace2f&)anisotropy.xform,
+          rotation.factor, rotation.map ? rotation.map->getIE() : nullptr, (const ispc::AffineSpace2f&)rotation.xform,
+          normal.factor, normal.map ? normal.map->getIE() : nullptr, (const ispc::AffineSpace2f&)normal.xform, (const ispc::LinearSpace2f&)normal.rot,
+
+          coat.factor, coat.map ? coat.map->getIE() : nullptr, (const ispc::AffineSpace2f&)coat.xform,
+          coatIor.factor, coatIor.map ? coatIor.map->getIE() : nullptr, (const ispc::AffineSpace2f&)coatIor.xform,
+          (const ispc::vec3f&)coatColor.factor, coatColor.map ? coatColor.map->getIE() : nullptr, (const ispc::AffineSpace2f&)coatColor.xform,
+          coatThickness.factor, coatThickness.map ? coatThickness.map->getIE() : nullptr, (const ispc::AffineSpace2f&)coatThickness.xform,
+          coatRoughness.factor, coatRoughness.map ? coatRoughness.map->getIE() : nullptr, (const ispc::AffineSpace2f&)coatRoughness.xform,
+          coatNormal.factor, coatNormal.map ? coatNormal.map->getIE() : nullptr, (const ispc::AffineSpace2f&)coatNormal.xform, (const ispc::LinearSpace2f&)coatNormal.rot,
+
+          sheen.factor, sheen.map ? sheen.map->getIE() : nullptr, (const ispc::AffineSpace2f&)sheen.xform,
+          (const ispc::vec3f&)sheenColor.factor, sheenColor.map ? sheenColor.map->getIE() : nullptr, (const ispc::AffineSpace2f&)sheenColor.xform,
+          sheenRoughness.factor, sheenRoughness.map ? sheenRoughness.map->getIE() : nullptr, (const ispc::AffineSpace2f&)sheenRoughness.xform,
+
+          opacity.factor, opacity.map ? opacity.map->getIE() : nullptr, (const ispc::AffineSpace2f&)opacity.xform,
+
+          thin,
+          backlight.factor, backlight.map ? backlight.map->getIE() : nullptr, (const ispc::AffineSpace2f&)backlight.xform,
+          thickness.factor, thickness.map ? thickness.map->getIE() : nullptr, (const ispc::AffineSpace2f&)thickness.xform,
+          
+          outsideIor,
+          (const ispc::vec3f&)outsideTransmissionColor,
+          outsideTransmissionDepth);
       }
     };