+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Version History
 ---------------
 
+### Changes in v1.6.0:
+
+-   Added Principled and CarPaint materials
+-   Improved flexibility of the tone mapper
+-   Added ability to configure Embree scene flags via OSPModel parameters
+
 ### Changes in v1.5.0:
 
 -   TetrahedralVolume now generalized to take both tet and hex data, now called